move device link to it's own namespace
